Add files via upload
A face recognition attendance system works by capturing and analyzing a person's face to identify and verify them, and then comparing that information to a database of previously collected data: 
1. Face detection
The system locates the person's face in real-time. 




2. Data transformation
The system converts the captured facial information into a set of data or vectors based on the person's facial features. 
3. Face match
The system compares the data to the information in the database to verify the person's identity. 
4. Attendance recorded
If the image matches the data in the database, the system records the person's attendance. 
Face recognition attendance systems can help prevent buddy-clocking and ensure accuracy and consistency in the calculation of hours worked. 
Some factors that can affect the accuracy of face recognition attendance systems include:
Lighting: The person's face should be well lit.
Background: There should not be any multiple people or objects in the background.